moves enums to sub-package constants
