+++ src/main/java/server/response/ResponseBuilder.java
@@ -1,5 +1,7 @@
 package server.response;
 
+import server.constants.StatusCode;
+
 public class ResponseBuilder {
     public Response build(StatusCode statusCode, String headerName, String headerValue, String body) {
         return new Response()
