+++ src/test/java/server/routing/RoutesTest.java
@@ -1,7 +1,8 @@
 package server.routing;
 
 import org.junit.*;
-import server.request.Method;
+import server.constants.StatusCode;
+import server.constants.Method;
 import server.response.*;
 
 import java.util.*;
